@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
 class Person extends StatelessWidget {
-  final Map<String, dynamic> person;
-  final Map<String, Color> _colors = {
+  final Map<String, dynamic>? person;
+  final Map<String, Color>? _colors = {
     "nice": Color.fromRGBO(220, 255, 220, 1.0),
     "naughty": Color.fromRGBO(255, 220, 220, 1.0)
   };
@@ -10,10 +10,10 @@ class Person extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    Color color = _colors[person['status']];
+    Color color = _colors![person!['status']]!;
 
     return Container(
-      child: Text("${person['first']} ${person['last']}",
+      child: Text("${person!['first']} ${person!['last']}",
           style: TextStyle(
             fontWeight: FontWeight.bold,
             fontSize: 20,
